refactor: login.kakao 패키지 만들고 KakaoLogin 관련 분리해냄
필요없는 AuthenticationToken 클래스 삭제함
