+++ api/src/main/kotlin/team/guin/account/AccountApiService.kt
@@ -4,22 +4,11 @@ import org.springframework.data.repository.findByIdOrNull
 import org.springframework.stereotype.Service
 import org.springframework.transaction.annotation.Transactional
 import team.guin.domain.account.Account
-import team.guin.security.kakao.AccountProfile
-import team.guin.security.kakao.KakaoApiService
 
 @Service
 class AccountApiService(
     private val accountApiRepository: AccountApiRepository,
-    private val kakaoApiService: KakaoApiService,
 ) {
-    fun joinOrLogin(accessToken: String): AccountProfile {
-        val kakaoDetailProfile = kakaoApiService.fetchKakaoDetailProfile(accessToken)
-
-        val account = accountApiRepository.findByEmail(kakaoDetailProfile.email)
-            ?: accountApiRepository.save(kakaoDetailProfile.toEntity())
-
-        return AccountProfile.from(account)
-    }
 
     @Transactional
     fun updateInfo(id: Long, nickname: String, imageId: String): Account {
