--- api/src/main/kotlin/team/guin/account/dto/AccountJoinRequest.kt
@@ -1,3 +1,0 @@
-package team.guin.account.dto
-
-class AccountJoinRequest(val accessToken: String)
