+++ api/src/main/kotlin/team/guin/login/kakao/dto/KakaoLoginRequest.kt
@@ -1,0 +1,3 @@
+package team.guin.login.kakao.dto
+
+class KakaoLoginRequest(val accessToken: String)
